requirements.txt
setuptools~=59.6.0
Twisted~=21.7.0
PyYAML~=6.0
pandas~=1.3.5
MetaTrader5~=5.0.36
pytz~=2021.3
jsons~=1.6.2
pid~=3.0.4